Feature/exception handling

+++ src/main/kotlin/com/teamsparta/todoapp/domain/cards/service/CardServiceImpl.kt
@@ -1,6 +1,5 @@
 package com.teamsparta.todoapp.domain.cards.service
 
-import com.teamsparta.todoapp.domain.cards.dto.CardAndCommentResponse
 import com.teamsparta.todoapp.domain.cards.dto.CardResponse
 import com.teamsparta.todoapp.domain.cards.dto.CreateCardRequest
 import com.teamsparta.todoapp.domain.cards.dto.UpdateCardRequest
@@ -11,66 +10,73 @@ import com.teamsparta.todoapp.domain.exception.ModelNotFoundException
 import jakarta.transaction.Transactional
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
-import java.time.OffsetDateTime
-import com.teamsparta.todoapp.domain.comment.service.CommentService
+import com.teamsparta.todoapp.domain.user.model.UserEntity
+import com.teamsparta.todoapp.domain.user.repository.UserRepository
+import org.springframework.security.access.AccessDeniedException
 
 @Service
 
 class CardServiceImpl(
-    private val cardRepository: CardRepository, private val commentService: CommentService
+    private val cardRepository: CardRepository, private val userRepository: UserRepository
 ) : CardService {
 
     override fun getAllCardList(): List<CardResponse> {
         return cardRepository.findAll().map { it.toResponse() }
     }
 
-    override fun getCommentsByCardId(cardId: Long): CardAndCommentResponse {
-        val cardResponse = cardRepository.getCardById(cardId).toResponse()
-        val comments = commentService.getCommentsByCardId(cardId)
-
-        return CardAndCommentResponse(
-            card = cardResponse, comments = comments
-        )
-    }
-
     override fun getCardById(cardId: Long): CardResponse {
         val card = cardRepository.findByIdOrNull(cardId) ?: throw ModelNotFoundException("Card", cardId)
         return card.toResponse()
+
     }
 
     @Transactional
-    override fun createCard(request: CreateCardRequest): CardResponse {
-        return cardRepository.save(
+    override fun createCard(request: CreateCardRequest,userId: Long): CardResponse {
+        val user : UserEntity = userRepository.findByIdOrNull(userId) ?: throw ModelNotFoundException("User", userId)
+        val savedCard = cardRepository.save(
             Card(
                 title = request.title,
                 description = request.description,
-                writer = request.writer,
-                createdAt = OffsetDateTime.now(),
+                user = user
             )
-        ).toResponse()
+        )
+
+        return savedCard.toResponse()
     }
 
-    @Transactional
-    override fun updateCard(cardId: Long, request: UpdateCardRequest): CardResponse {
-        val card = cardRepository.findByIdOrNull(cardId) ?: throw ModelNotFoundException("Card", cardId)
-        val (title, description, writer) = request
 
+
+    @Transactional
+    override fun updateCard(cardId: Long, userId: Long, request: UpdateCardRequest): CardResponse {
+        val card = cardRepository.findByIdOrNull(cardId) ?: throw ModelNotFoundException("Card", cardId)
+        //카드 작성 유저와 현재 유저가 같은지 검증과정
+        if (card.user.id != userId) {
+            throw AccessDeniedException("You do not have permission to update this card.")
+            //AccessDeniedException은 customAccessDeniedHandler에서 따로 조절
+        }
+        val (title, description) = request
         card.title = title
         card.description = description
-        card.writer = writer
 
-        return cardRepository.save(card).toResponse()
+        val updatedCard = cardRepository.save(card)
+        return updatedCard.toResponse()
     }
 
     @Transactional
-    override fun deleteCard(cardId: Long) {
+    override fun deleteCard(cardId: Long,userId: Long) {
         val card = cardRepository.findByIdOrNull(cardId) ?: throw ModelNotFoundException("Card", cardId)
+        if (card.user.id != userId) {
+            throw AccessDeniedException("You do not have permission to delete this card.")
+        }
         cardRepository.delete(card)
     }
 
     @Transactional
-    override fun toggleCardCompletion(cardId: Long): CardResponse {
+    override fun toggleCardCompletion(cardId: Long,userId: Long): CardResponse {
         val card = cardRepository.findByIdOrNull(cardId) ?: throw ModelNotFoundException("Card", cardId)
+        if (card.user.id != userId) {
+            throw AccessDeniedException("You do not have permission to toggle this card.")
+        }
         card.completed = !card.completed
         return card.toResponse()
     }
